re-organise the files and direcotires
